'delete'

+++ src/main/java/com/api_vendinha/api/controller/VendaController.java
@@ -52,7 +52,7 @@ public class VendaController {
     }
 
     @DeleteMapping("/deletar/{id}")
-    public VendaResponseDto deletar(@PathVariable Long id) {
+    public VendaResponseDto deletar(@PathVariable Integer id) {
         return vendaServiceInterface.deletar(id);
 
     }
